Empty statment added

+++ src/edu/mit/compilers/ll/LlLiteralInt.java
@@ -4,18 +4,18 @@ package edu.mit.compilers.ll;
  * Created by devinmorgan on 11/18/16.
  */
 public class LlLiteralInt extends LlLiteral {
-    private final int intValue;
+    private final long intValue;
 
-    public LlLiteralInt(int intValue) {
+    public LlLiteralInt(long intValue) {
         this.intValue = intValue;
     }
 
-    public int getIntValue() {
+    public long getIntValue() {
         return this.intValue;
     }
 
     @Override
     public String toString() {
-        return Integer.toString(this.intValue);
+        return Long.toString(this.intValue);
     }
 }
